1. diet native getter, setter on Object 2. diet ByteCode 3. use get, setIndexedProperty in ByteCodeInterpreter 4. implement get, setIndexedProperty for typed array 5. move implemention of CallNativeFunctionOpcode to FunctionObject::call 6. remove m_contentLength, m_has8BitContent in RopeString 7. optimize ByteCodeInterpreter::getObjectPrecomputedCaseOperation, setObjectPrecomputedCaseOperation 8. diet Arguments object

+++ src/interpreter/ByteCode.h
@@ -299,7 +299,7 @@ public:
     {
     }
     ByteCodeRegisterIndex m_registerIndex;
-    size_t m_index;
+    uint16_t m_index;
 
 #ifndef NDEBUG
     virtual void dump()
@@ -318,7 +318,7 @@ public:
     {
     }
     ByteCodeRegisterIndex m_registerIndex;
-    size_t m_index;
+    uint16_t m_index;
 
 #ifndef NDEBUG
     virtual void dump()
@@ -338,7 +338,7 @@ public:
     {
     }
     ByteCodeRegisterIndex m_registerIndex;
-    size_t m_upperIndex;
+    uint16_t m_upperIndex;
     size_t m_index;
 
 #ifndef NDEBUG
@@ -359,7 +359,7 @@ public:
     {
     }
     ByteCodeRegisterIndex m_registerIndex;
-    size_t m_upperIndex;
+    uint16_t m_upperIndex;
     size_t m_index;
 
 #ifndef NDEBUG
@@ -450,7 +450,7 @@ public:
     }
 
     ByteCodeRegisterIndex m_registerIndex;
-    size_t m_argumentCount;
+    uint16_t m_argumentCount;
 #ifndef NDEBUG
     virtual void dump()
     {
